+++ app/javascript/application.js
@@ -1,5 +1,5 @@
 import "bootstrap";
 import "controllers/stats_chart";
-import "controllers/shop_preview";
+import "controllers/deal_countdown";
 import "controllers/push_notifications";
 import "controllers/weekly_stats_chart";